+++ Sources/TuistDependencies/Mappers/ExternalProjectsPlatformNarrowerGraphMapper.swift
@@ -24,21 +24,55 @@ public struct ExternalProjectsPlatformNarrowerGraphMapper: GraphMapping {
         graph.targets = Dictionary(uniqueKeysWithValues: graph.targets.map { projectPath, projectTargets in
             let project = graph.projects[projectPath]!
             let projectTargets = Dictionary(uniqueKeysWithValues: projectTargets.map { targetName, target in
-                let graphTarget = GraphTarget(path: projectPath, target: target, project: project)
-                var target = target
-
-                /**
-                 We only include the destinations whose platform is included in the list of the target supported platforms.
-                 */
-                if let targetFilteredPlatforms = externalTargetSupportedPlatforms[graphTarget] {
-                    target.destinations = target.destinations.filter { destination in
-                        targetFilteredPlatforms.contains(destination.platform)
-                    }
-                }
-                return (targetName, target)
+                (
+                    targetName,
+                    mapTarget(
+                        target: target,
+                        project: project,
+                        externalTargetSupportedPlatforms: externalTargetSupportedPlatforms
+                    )
+                )
             })
             return (projectPath, projectTargets)
         })
+        graph.projects = Dictionary(uniqueKeysWithValues: graph.projects.map { projectPath, project in
+            var project = project
+            project.targets = project.targets.map { mapTarget(
+                target: $0,
+                project: project,
+                externalTargetSupportedPlatforms: externalTargetSupportedPlatforms
+            ) }
+            return (projectPath, project)
+        })
+
         return (graph, [])
+    }
+
+    private func mapTarget(
+        target: Target,
+        project: Project,
+        externalTargetSupportedPlatforms: [GraphTarget: Set<Platform>]
+    ) -> Target {
+        /**
+         We only include the destinations whose platform is included in the list of the target supported platforms.
+         */
+        var target = target
+        let graphTarget = GraphTarget(path: project.path, target: target, project: project)
+        if project.isExternal, let targetFilteredPlatforms = externalTargetSupportedPlatforms[graphTarget] {
+            target.destinations = target.destinations.filter { destination in
+                targetFilteredPlatforms.contains(destination.platform)
+            }
+
+            // By changing the destinations we also need to adapt the deployment targets accordingly to account for possibly
+            // removed destinations
+            target.deploymentTargets = .init(
+                iOS: targetFilteredPlatforms.contains(.iOS) ? target.deploymentTargets.iOS : nil,
+                macOS: targetFilteredPlatforms.contains(.macOS) ? target.deploymentTargets.macOS : nil,
+                watchOS: targetFilteredPlatforms.contains(.watchOS) ? target.deploymentTargets.watchOS : nil,
+                tvOS: targetFilteredPlatforms.contains(.tvOS) ? target.deploymentTargets.tvOS : nil,
+                visionOS: targetFilteredPlatforms.contains(.visionOS) ? target.deploymentTargets.visionOS : nil
+            )
+        }
+        return target
     }
 }