+++ Tests/TuistAcceptanceTests/ListTargetsAcceptanceTests.swift
@@ -7,7 +7,7 @@ import XCTest
 
 final class ListTargetsAcceptanceTestiOSWorkspaceWithMicrofeatureArchitecture: TuistAcceptanceTestCase {
     func test_ios_workspace_with_microfeature_architecture() async throws {
-        try setUpFixture("ios_workspace_with_microfeature_architecture")
+        try setUpFixture(.iosWorkspaceWithMicrofeatureArchitecture)
         try await run(GenerateCommand.self)
         try listTargets(for: "UIComponents")
         try listTargets(for: "Core")